requirements.txt
streamlit
tensorflow
Pillow
gdown
